First version of confirmationui for Trusty passing the VTS test
confirmationui: trusty ipc for large messages

This patch implements a simple synchronous ipc protocol for long
message/response transactions. In this context "long" meas messages that
are to long for a single message buffer. The event loop handles only one
connection at a time. This reflects the fact that confirmationui depends
on a limited resource (the screen) and can only have one active session
at a time.

It also uses the generic operation and message formatting to implement a
minimal confirmationui implementation passing VTS test. Alas it still
misses integration with a secure frame buffer.

Bug: 146078942
Test: VtsHalConfirmationUIV1_0TargetTest
Change-Id: Ic4f104de6cad2f4d11d753dd28a1e00781889d2a